Operações DELETE e UPDATE implementadas

+++ src/main/java/com/veterinaria/demo/model/dto/animal/AnimalRequestDTO.java
@@ -21,7 +21,7 @@ public record AnimalRequestDTO(
 
         @Positive
         @NotNull
-        int age,
+        Long age,
 
         @NotNull
         Long tutorId
